Fixed all of my powershell scripts; am on so much of an automation kick that I may make it possible to completely update my dataset just by running one command, this is exciting.

+++ random.ps1
@@ -1,10 +1,11 @@
-#cd images
-#Get-ChildItem | ForEach-Object {
-#	$targetFolder = $_
-#	cd $targetFolder
-$d = gci ".\images\*.jpg" | Where {$_.psIsContainer -eq $false} | resolve-path
-Copy-Item $d -destination .\images\
-Remove-Item $d
-$d = gci ".\images\*.jpg" | Where {$_.psIsContainer -eq $false} | resolve-path  |  get-random -count 2
-Copy-Item $d  -destination .\images\selected
-Remove-Item $d
+gci -Path ".\directory_names\*" | Where {$_.psIsContainer -eq $true} | ForEach-Object {
+	$targetFolder = $_
+    $d = gci -Path (".\images\test_" + $targetFolder.Name + "\*.jpg") | Where {$_.psIsContainer -eq $false} | resolve-path
+    if ($d)  {
+        Copy-Item $d -destination (".\images\train_" + $targetFolder.Name + "\")
+        Remove-Item $d
+    }
+    $d = gci -Path (".\images\train_" + $targetFolder.Name + "\*.jpg") | Where {$_.psIsContainer -eq $false} | resolve-path  |  get-random -count 20
+    Copy-Item $d -destination (".\images\test_" + $targetFolder.Name + "\")
+    Remove-Item $d
+}
